Product Wizard: Update default
Previously both options were being selected as a default, most people won't have much a need to do both types at one time so that's not the best default action.
It's not great to assume the users intentions, doing so would only increase the needed clicks to get stuff done; that in mind, a sane default of none selected is now in place.